+++ cmd/cored/main.go
@@ -3,9 +3,10 @@ package main
 import (
 	"os"
 
-	"github.com/CoreumFoundation/coreum/cored/app"
 	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
 	"github.com/ignite-hq/cli/ignite/pkg/cosmoscmd"
+
+	"github.com/CoreumFoundation/coreum/app"
 )
 
 func main() {
