+++ testutil/network/network.go
@@ -18,7 +18,7 @@ import (
 	tmrand "github.com/tendermint/tendermint/libs/rand"
 	tmdb "github.com/tendermint/tm-db"
 
-	"github.com/CoreumFoundation/coreum/cored/app"
+	"github.com/CoreumFoundation/coreum/app"
 )
 
 type (
